+++ ToolManagementSystem.Shared/Models/CycleItem.cs
@@ -5,11 +5,11 @@ namespace ToolManagementSystem.Shared.Models
 {
     public partial class CycleItem
     {
-        public long CycleItemId { get; set; }
-        public long? PrevToolStatusId { get; set; }
-        public long CurrToolStatusId { get; set; }
-        public long? NextToolStatusId { get; set; }
-        public long CycleId { get; set; }
+        public int CycleItemId { get; set; }
+        public int? PrevToolStatusId { get; set; }
+        public int CurrToolStatusId { get; set; }
+        public int? NextToolStatusId { get; set; }
+        public int CycleId { get; set; }
 
         public virtual Cycle Cycle { get; set; }
     }
