+++ ToolManagementSystem.Shared/Models/OrderDocument.cs
@@ -5,9 +5,9 @@ namespace ToolManagementSystem.Shared.Models
 {
     public partial class OrderDocument
     {
-        public long OrderDocumentId { get; set; }
-        public long OrderId { get; set; }
-        public long DocumentId { get; set; }
+        public int OrderDocumentId { get; set; }
+        public int OrderId { get; set; }
+        public int DocumentId { get; set; }
 
         public virtual Document Document { get; set; }
         public virtual Order Order { get; set; }
