+++ ToolManagementSystem.Shared/Models/OrderTool.cs
@@ -10,9 +10,9 @@ namespace ToolManagementSystem.Shared.Models
             OrderToolOperatingTime = new HashSet<OrderToolOperatingTime>();
         }
 
-        public long OrderToolId { get; set; }
-        public long OrderId { get; set; }
-        public long ToolId { get; set; }
+        public int OrderToolId { get; set; }
+        public int OrderId { get; set; }
+        public int ToolId { get; set; }
         public DateTime ArrivedDate { get; set; }
         public DateTime? LeftDate { get; set; }
