+++ ToolManagementSystem.Shared/Models/Permission.cs
@@ -11,12 +11,12 @@ namespace ToolManagementSystem.Shared.Models
             RolePermission = new HashSet<RolePermission>();
         }
 
-        public long PermissionId { get; set; }
+        public int PermissionId { get; set; }
         public string Name { get; set; }
         public string ViewName { get; set; }
         public string Alias { get; set; }
         public int? OrderNo { get; set; }
-        public long? ParentPermissionId { get; set; }
+        public int? ParentPermissionId { get; set; }
 
         public virtual Permission ParentPermission { get; set; }
         public virtual ICollection<Permission> InverseParentPermission { get; set; }
