+++ ToolManagementSystem.Shared/Models/ReportHistory.cs
@@ -5,11 +5,11 @@ namespace ToolManagementSystem.Shared.Models
 {
     public partial class ReportHistory
     {
-        public long ReportHistoryId { get; set; }
+        public int ReportHistoryId { get; set; }
         public string Name { get; set; }
-        public long CreatorId { get; set; }
+        public int CreatorId { get; set; }
         public DateTime CreationDate { get; set; }
-        public long ReportId { get; set; }
+        public int ReportId { get; set; }
 
         public virtual User Creator { get; set; }
         public virtual Report Report { get; set; }
