+++ ToolManagementSystem.Shared/Models/RolePermissionRight.cs
@@ -6,12 +6,12 @@ namespace ToolManagementSystem.Shared.Models
     public partial class RolePermissionRight
     {
         public long RoleRightPermission { get; set; }
-        public long RolePermissionId { get; set; }
-        public long RightId { get; set; }
+        public int RolePermissionId { get; set; }
+        public int RightId { get; set; }
         public DateTime CreationDate { get; set; }
-        public long CreatorId { get; set; }
+        public int CreatorId { get; set; }
         public DateTime? LastUpdate { get; set; }
-        public long? LastUpdatorId { get; set; }
+        public int? LastUpdatorId { get; set; }
 
         public virtual User Creator { get; set; }
         public virtual User LastUpdator { get; set; }
