+++ ToolManagementSystem.Shared/Models/ToolHistory.cs
@@ -5,11 +5,11 @@ namespace ToolManagementSystem.Shared.Models
 {
     public partial class ToolHistory
     {
-        public long ToolHistoryId { get; set; }
-        public long ToolId { get; set; }
-        public long ToolStatusId { get; set; }
+        public int ToolHistoryId { get; set; }
+        public int ToolId { get; set; }
+        public int ToolStatusId { get; set; }
         public DateTime CreationDate { get; set; }
-        public long CreatorId { get; set; }
+        public int CreatorId { get; set; }
         public string Message { get; set; }
         public string Commentary { get; set; }
