+++ ToolManagementSystem.Shared/Models/UserRole.cs
@@ -5,13 +5,13 @@ namespace ToolManagementSystem.Shared.Models
 {
     public partial class UserRole
     {
-        public long UserRoleId { get; set; }
-        public long UserId { get; set; }
-        public long RoleId { get; set; }
+        public int UserRoleId { get; set; }
+        public int UserId { get; set; }
+        public int RoleId { get; set; }
         public DateTime CreationDate { get; set; }
-        public long CreatorId { get; set; }
+        public int CreatorId { get; set; }
         public DateTime? LastUpdate { get; set; }
-        public long? LastUpdatorId { get; set; }
+        public int? LastUpdatorId { get; set; }
 
         public virtual User Creator { get; set; }
         public virtual User LastUpdator { get; set; }
