+++ ToolManagementSystem.Shared/Models/WorkOrder.cs
@@ -12,14 +12,14 @@ namespace ToolManagementSystem.Shared.Models
             WorkOrderTool = new HashSet<WorkOrderTool>();
         }
 
-        public long WorkOrderId { get; set; }
+        public int WorkOrderId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? EndDate { get; set; }
         public DateTime EstimatedEndDate { get; set; }
-        public long CreatorId { get; set; }
-        public long WorkOrderStatusId { get; set; }
+        public int CreatorId { get; set; }
+        public int WorkOrderStatusId { get; set; }
         public string Name { get; set; }
-        public long? ResponsibleUserId { get; set; }
+        public int? ResponsibleUserId { get; set; }
 
         public virtual User Creator { get; set; }
         public virtual User ResponsibleUser { get; set; }
