+++ ToolManagementSystem.Shared/Models/WorkOrderDocument.cs
@@ -5,9 +5,9 @@ namespace ToolManagementSystem.Shared.Models
 {
     public partial class WorkOrderDocument
     {
-        public long WorkOrderDocumentId { get; set; }
-        public long WorkOrderId { get; set; }
-        public long DocumentId { get; set; }
+        public int WorkOrderDocumentId { get; set; }
+        public int WorkOrderId { get; set; }
+        public int DocumentId { get; set; }
 
         public virtual Document Document { get; set; }
         public virtual WorkOrder WorkOrder { get; set; }
