+++ ToolManagementSystem.Shared/Models/WorkOrderHistory.cs
@@ -5,12 +5,12 @@ namespace ToolManagementSystem.Shared.Models
 {
     public partial class WorkOrderHistory
     {
-        public long WorkOrderHistoryId { get; set; }
-        public long WorkOrderId { get; set; }
+        public int WorkOrderHistoryId { get; set; }
+        public int WorkOrderId { get; set; }
         public DateTime CreationDate { get; set; }
-        public long CreatorId { get; set; }
+        public int CreatorId { get; set; }
         public string Message { get; set; }
-        public long WorkOrderStatusId { get; set; }
+        public int WorkOrderStatusId { get; set; }
 
         public virtual User Creator { get; set; }
         public virtual WorkOrder WorkOrder { get; set; }
